code/requirements2.txt
absl-py==0.15.0
adal==1.2.7
applicationinsights==0.11.10
astunparse==1.6.3
attrs==21.2.0
autopep8==1.6.0
azure-common==1.1.27
azure-core==1.21.0
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.36.1
azureml-core==1.36.0.post2
azureml-dataprep==2.24.4
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.0.3
azureml-dataset-runtime==1.36.0
azureml-pipeline==1.36.0
azureml-pipeline-core==1.36.0
azureml-pipeline-steps==1.36.0
azureml-sdk==1.36.0
azureml-telemetry==1.36.0
azureml-train-automl-client==1.36.0
azureml-train-core==1.36.0
azureml-train-restclients-hyperdrive==1.36.0
backports.tempfile==1.0
backports.weakref==1.0.post1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudpickle==1.6.0
contextlib2==21.6.0
cryptography==3.4.8
cycler==0.11.0
decorator==5.1.0
distro==1.6.0
dm-tree==0.1.6
docker==5.0.3
dotnetcore2==2.1.21
flatbuffers==2.0
fusepy==3.0.1
gast==0.4.0
gin-config==0.5.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
gym==0.21.0
h5py==3.5.0
idna==3.3
iniconfig==1.1.1
isodate==0.6.0
jeepney==0.7.1
jmespath==0.10.0
jsonpickle==2.0.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
matplotlib==3.4.3
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
pathspec==0.9.0
Pillow==8.4.0
pluggy==1.0.0
portalocker==1.7.1
protobuf==3.19.1
py==1.11.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
SecretStorage==3.3.1
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tfp-nightly==0.15.0.dev20211109
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.3
